data_collection/requirements-dev.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile requirements-dev.in
#
aiohttp==3.8.1
    # via slackclient
aiosignal==1.2.0
    # via aiohttp
async-timeout==4.0.1
    # via aiohttp
asynctest==0.13.0
    # via aiohttp
attrs==21.2.0
    # via
    #   aiohttp
    #   automat
    #   jsonschema
    #   service-identity
    #   twisted
automat==20.2.0
    # via twisted
awscli==1.22.24
    # via -r requirements.in
backports.entry-points-selectable==1.1.1
    # via virtualenv
backports.zoneinfo==0.2.1
    # via
    #   pytz-deprecation-shim
    #   tzlocal
black==21.12b0
    # via -r requirements-dev.in
boto==2.49.0
    # via spidermon
boto3==1.20.24
    # via
    #   -r requirements.in
    #   spidermon
botocore==1.23.24
    # via
    #   awscli
    #   boto3
    #   s3transfer
cachetools==4.2.4
    # via premailer
certifi==2021.10.8
    # via
    #   requests
    #   sentry-sdk
cffi==1.15.0
    # via cryptography
cfgv==3.3.1
    # via pre-commit
charset-normalizer==2.0.9
    # via
    #   aiohttp
    #   requests
chompjs==1.1.6
    # via -r requirements.in
click==8.0.3
    # via
    #   -r requirements.in
    #   black
    #   pip-tools
colorama==0.4.3
    # via awscli
constantly==15.1.0
    # via twisted
cryptography==36.0.0
    # via
    #   pyopenssl
    #   scrapy
    #   service-identity
cssselect==1.1.0
    # via
    #   parsel
    #   premailer
    #   scrapy
cssutils==2.3.0
    # via premailer
dateparser==1.1.0
    # via -r requirements.in
distlib==0.3.4
    # via virtualenv
docutils==0.15.2
    # via awscli
filelock==3.4.0
    # via virtualenv
flake8==4.0.1
    # via -r requirements-dev.in
frozenlist==1.2.0
    # via
    #   aiohttp
    #   aiosignal
greenlet==1.1.2
    # via sqlalchemy
h2==3.2.0
    # via
    #   scrapy
    #   twisted
hpack==3.0.0
    # via h2
hyperframe==5.2.0
    # via h2
hyperlink==21.0.0
    # via twisted
identify==2.4.0
    # via pre-commit
idna==3.3
    # via
    #   hyperlink
    #   jsonschema
    #   requests
    #   yarl
importlib-metadata==4.2.0
    # via
    #   backports.entry-points-selectable
    #   click
    #   cssutils
    #   flake8
    #   jsonschema
    #   pep517
    #   pre-commit
    #   sqlalchemy
    #   virtualenv
incremental==21.3.0
    # via twisted
isort==5.10.1
    # via -r requirements-dev.in
itemadapter==0.4.0
    # via
    #   -r requirements.in
    #   itemloaders
    #   scrapy
itemloaders==1.0.4
    # via scrapy
jinja2==3.0.3
    # via
    #   -r requirements.in
    #   spidermon
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
    #   itemloaders
jsonpointer==2.2
    # via jsonschema
jsonschema[format]==3.2.0
    # via spidermon
lxml==4.7.1
    # via
    #   parsel
    #   premailer
    #   scrapy
markupsafe==2.0.1
    # via jinja2
mccabe==0.6.1
    # via flake8
multidict==5.2.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==0.4.3
    # via black
nodeenv==1.6.0
    # via pre-commit
parsel==1.6.0
    # via
    #   itemloaders
    #   scrapy
pathspec==0.9.0
    # via black
pep517==0.12.0
    # via pip-tools
pip-tools==6.4.0
    # via -r requirements-dev.in
platformdirs==2.4.0
    # via
    #   black
    #   virtualenv
pre-commit==2.16.0
    # via -r requirements-dev.in
premailer==3.10.0
    # via spidermon
priority==1.3.0
    # via twisted
protego==0.1.16
    # via scrapy
psycopg2-binary==2.9.2
    # via -r requirements.in
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
    #   service-identity
pyasn1-modules==0.2.8
    # via service-identity
pycodestyle==2.8.0
    # via flake8
pycparser==2.21
    # via cffi
pydispatcher==2.0.5
    # via scrapy
pyflakes==2.4.0
    # via flake8
pyopenssl==21.0.0
    # via scrapy
pyrsistent==0.18.0
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   -r requirements.in
    #   botocore
    #   dateparser
python-decouple==3.5
    # via -r requirements.in
python-slugify==5.0.2
    # via spidermon
pytz==2021.3
    # via dateparser
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
pyyaml==5.4.1
    # via
    #   awscli
    #   pre-commit
queuelib==1.6.2
    # via scrapy
regex==2021.11.10
    # via dateparser
requests==2.26.0
    # via premailer
rfc3987==1.3.8
    # via jsonschema
rsa==4.7.2
    # via awscli
s3transfer==0.5.0
    # via
    #   awscli
    #   boto3
schematics==2.1.1
    # via -r requirements.in
scrapy==2.5.1
    # via
    #   -r requirements.in
    #   spidermon
sentry-sdk==1.5.1
    # via spidermon
service-identity==21.1.0
    # via scrapy
six==1.16.0
    # via
    #   automat
    #   jsonschema
    #   parsel
    #   protego
    #   pyopenssl
    #   python-dateutil
    #   service-identity
    #   spidermon
    #   virtualenv
    #   w3lib
slackclient==2.9.3
    # via spidermon
spidermon[monitoring]==1.15.2
    # via -r requirements.in
sqlalchemy==1.4.28
    # via -r requirements.in
strict-rfc3339==0.7
    # via jsonschema
text-unidecode==1.3
    # via python-slugify
toml==0.10.2
    # via pre-commit
tomli==1.2.3
    # via
    #   black
    #   pep517
twisted[http2]==21.7.0
    # via scrapy
typed-ast==1.5.1
    # via black
typing-extensions==4.0.1
    # via
    #   aiohttp
    #   async-timeout
    #   black
    #   importlib-metadata
    #   twisted
    #   yarl
tzdata==2021.5
    # via pytz-deprecation-shim
tzlocal==4.1
    # via dateparser
urllib3==1.26.7
    # via
    #   botocore
    #   requests
    #   sentry-sdk
virtualenv==20.10.0
    # via pre-commit
w3lib==1.22.0
    # via
    #   itemloaders
    #   parsel
    #   scrapy
webcolors==1.11.1
    # via jsonschema
wheel==0.37.0
    # via pip-tools
yarl==1.7.2
    # via aiohttp
zipp==3.6.0
    # via
    #   importlib-metadata
    #   pep517
zope.interface==5.4.0
    # via
    #   scrapy
    #   twisted

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
